bot/requirements.txt
discord.py
asyncpg
jishaku
psutil
nltk
lxml
bs4
xdice
oauth2client==4.1.3
google-api-python-client==1.8.3
git+https://github.com/Rapptz/discord-ext-menus